(PC-41763)[PRO] refactor: place phone util files in good directories
