+++ pro/src/pages/User/UserProfile/UserPhoneForm/validationSchema.ts
@@ -1,6 +1,6 @@
 import * as yup from 'yup'
 
-import { phoneNumberSchema } from '@/ui-kit/form/PhoneNumberInput/commons/phoneNumberSchema'
+import { phoneNumberSchema } from '@/commons/utils/yup/phoneNumberSchema'
 
 export const validationSchema = yup.object().shape({
   phoneNumber: phoneNumberSchema().required(
